Time Table working
Download button now works, image gets downloaded in the pictures folder.
NEEDED : app needs External write storage, the permission has to be granted from the settings menu! instead it should provide the dialog box in which user can select the permission. and if the permission is denied then the download button should be hided from the time table fragment.