Data is now saved offline for faster loading
this build saves data offline by preserving an instance in the memory for later use. whenever app is offline it still shows posts and other user data and thus also helps in faster loading of news posts and other stuff.
It still needs a working internet connection to fetch images of posts! will try to save them in cache too in the later builds.